requirements.txt
requests
web3
asyncio_throttle
ScanWatch
PyYAML
